缓存策略优化，重新定义缓存的清理，添加过期数据的清理
git-svn-id: http://svn.simba.taobao.com/svn/PW_WindFramework/laboratory/trunk@2624 18ba2127-5a84-46d4-baec-3457e417f034

+++ wind/cache/strategy/WindEacceleratorCache.php
@@ -42,8 +42,9 @@ class WindEacceleratorCache extends AbstractWindCache {
 	 * @see AbstractWindCache#clear()
 	 * @return boolean
 	 */
-	public function clear() {
+	public function clear($expireOnly = false) {
 		eaccelerator_gc();
+	    if ($expireOnly) return true;
 		$cacheKeys = eaccelerator_list_keys();
 		foreach ($cacheKeys as $key) {
 			$this->delete(substr($key['name'], 1));
